+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Fixes
 
 - Handle traces sampler exception ([#1040](https://github.com/getsentry/sentry-dart/pull/1040))
+- tracePropagationTargets ignores invalid Regex ([#1040](https://github.com/getsentry/sentry-dart/pull/1040))
 
 ### Features